Prepare repository for publication
