--- environment.yaml
@@ -1,27 +1,0 @@
-name: timeflux_rasr-env
-dependencies:
-    - python>=3.7,<3.8
-    - numpy>=1.16.4,<1.17
-    - scipy>=1.1
-    - pyzmq>=17.1.2
-    - coloredlogs>=10.0
-    - pytables>=3.4.4
-    - jsonschema
-    - pytest>=4.0.2
-    - sphinx>=1.8
-    - sphinx_rtd_theme
-    - python-graphviz
-    - scikit-learn==0.20.2
-    - pandas
-    - scipy
-    - matplotlib
-    - seaborn
-    # Pip dependencies
-    - pip
-    - pip:
-        - mne
-        - pyxdf
-        - pyriemann
-        # prefer tarball for timeflux so that versioneer does not fail to detect version
-        - git+https://github.com/timeflux/timeflux_dsp@9494f40#egg=timeflux_dsp
-        - git+https://github.com/timeflux/timeflux#egg=timeflux_dsp
