+++ pyproject.toml
@@ -1,0 +1,11 @@
+# Enable version inference
+[tool.setuptools_scm]
+local_scheme = "no-local-version"
+
+# Generate documentation shim
+[tool.docinit]
+
+# https://setuptools.readthedocs.io/en/latest/build_meta.html
+[build-system]
+requires = ["setuptools>=42", "wheel", "setuptools_scm[toml]>=3.4", "docinit"]
+build-backend = "setuptools.build_meta"
