+++ setup.cfg
@@ -1,0 +1,43 @@
+[metadata]
+name = timeflux-rasr
+description = Implementation of rASR filtering.
+long_description = file: README.rst
+author = Timeflux Team
+author-email = contact@timeflux.io
+license = MIT
+home-page = https://timeflux.io
+project_urls =
+    Documentation = https://doc.timeflux.io/projects/timeflux-rasr/
+    Source Code = https://github.com/timeflux/timeflux_rasr
+    Bug Tracker = https://github.com/timeflux/timeflux_rasr/issues
+classifier =
+    Development Status :: 4 - Beta
+    Environment :: Console
+    Intended Audience :: Developers
+    Intended Audience :: Science/Research
+    License :: OSI Approved :: MIT License
+    Operating System :: OS Independent
+    Programming Language :: Python
+keywords = timeflux
+
+[options]
+packages = find:
+install_requires =
+    timeflux>=0.10.0
+    pyriemann@git+https://github.com/pyRiemann/pyRiemann
+
+[options.extras_require]
+dev =
+    pytest>=5.3
+    sphinx>=2.2
+    sphinx_rtd_theme>=0.4
+    setuptools_scm
+    docinit
+
+[docinit]
+name = Timeflux rASR
+parent_url = https://doc.timeflux.io
+logo_url = https://github.com/timeflux/timeflux/raw/master/doc/static/img/logo.png
+
+[build_sphinx]
+warning-is-error = 1
