+++ setup.py
@@ -1,0 +1,3 @@
+# Required for editable installs
+# https://discuss.python.org/t/specification-of-editable-installation/1564/
+import setuptools; setuptools.setup()
