--- utils/__init__.py
@@ -1,5 +1,0 @@
-__version__ = '0.1.0'
-
-from . import utils
-from . import config
-from . import viz
